Added a check so that you can't remove a user when he's an active teacher

+++ dokeos/user/lib/user_manager/component/deleter.class.php
@@ -29,6 +29,7 @@ class UserManagerDeleterComponent extends UserManagerComponent
 				exit;
 			}
 
+            if(UserDataManager :: get_instance()->user_deletion_allowed($user))
 			$success = $user->delete();
 
 			if($success)
